+++ src/content/collection.js
@@ -1,8 +1,17 @@
-// ※init
+
+// --------- Initialization
+
 setTimeout(() => {AutoCollectCollectibles(5);}, 5000);
 
-// ※functions
+// --------- Functions
+
 // Collect collectibles automatically.
+/**
+ * Automatically finds and clicks on the first collectible image element on the page
+ * and then reloads the page after a specified number of minutes.
+ *
+ * @param {number} [refreshMinute=5] - The number of minutes to wait before reloading the page. Defaults to 5 minutes.
+ */
 function AutoCollectCollectibles(refreshMinute = 5) {
     const c = [...document.getElementsByTagName('img')];
