+++ src/modular/rarity.js
